requirements.txt
pandas
requests
biopython